# auto_generate_video_fold6/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Auto Video Generation System API
  description: |
    ## Overview
    The Auto Video Generation System is a comprehensive AI-powered platform for creating viral short-form videos automatically. 
    
    This API allows you to:
    - Generate engaging video scripts using AI
    - Create stunning visuals with AI image generation
    - Synthesize natural-sounding voiceovers
    - Analyze trending topics and keywords
    - Schedule and publish content across multiple social platforms
    
    ## Authentication
    All API endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer YOUR_JWT_TOKEN
    ```
    
    ## Rate Limits
    - Free tier: 60 requests per minute
    - Pro tier: 300 requests per minute
    - Enterprise: Custom limits
    
    ## SDKs and Libraries
    Official SDKs are available for:
    - Python
    - JavaScript/TypeScript
    - React/Next.js
    
  version: "1.0.0"
  contact:
    name: Auto Video Generation Team
    email: support@autovideo.com
    url: https://docs.autovideo.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://autovideo.com/terms
servers:
  - url: https://api.autovideo.com/v1
    description: Production server
  - url: https://staging-api.autovideo.com/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /ai/text/generate-script:
    post:
      tags: [AI - Text Generation]
      summary: Generate video script
      description: |
        Generate an engaging video script based on topic and style preferences.
        
        **Features:**
        - Multiple writing styles (engaging, informative, humorous, professional)
        - Customizable duration and target audience
        - Keyword integration for SEO optimization
        - Real-time generation with quality scoring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptGenerationRequest'
            example:
              topic: "5 AI Tools That Will Change Your Life"
              style: "engaging"
              duration_seconds: 60
              target_audience: "tech enthusiasts"
              keywords: ["AI", "productivity", "automation"]
              tone: "enthusiastic"
      responses:
        '200':
          description: Script generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptGenerationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimit'
  /ai/text/generate-titles:
    post:
      tags: [AI - Text Generation]
      summary: Generate catchy titles
      description: Generate multiple engaging titles based on script content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TitleGenerationRequest'
      responses:
        '200':
          description: Titles generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitleGenerationResponse'

  # AI Image Generation Endpoints
  /ai/images/generate:
    post:
      tags: [AI - Image Generation]
      summary: Generate image from text
      description: |
        Create stunning visuals using AI image generation.
        
        **Supported Features:**
        - Multiple artistic styles (modern, vintage, minimalist, etc.)
        - Various aspect ratios (9:16, 16:9, 1:1, 4:3)
        - High-resolution output (720p, 1080p, 4K)
        - Negative prompts for quality control
        - Seed-based reproducibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageGenerationRequest'
            example:
              prompt: "Modern workspace with AI hologram, clean design, blue accents"
              style: "modern"
              aspect_ratio: "9:16"
              resolution: "1080p"
              negative_prompt: "blurry, low quality, distorted"
      responses:
        '200':
          description: Image generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageGenerationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /ai/images/generate-variations:
    post:
      tags: [AI - Image Generation]
      summary: Generate image variations
      description: Create multiple variations of an existing image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageVariationRequest'
      responses:
        '200':
          description: Variations generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageVariationResponse'

  # AI Audio Processing Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ScriptGenerationRequest:
      type: object
      required: [topic]
      properties:
        topic:
          type: string
          description: Main topic for the video script
          example: "5 AI Tools That Will Change Your Life"
        style:
          type: string
          enum: [engaging, informative, humorous, professional]
          default: engaging
        duration_seconds:
          type: integer
          minimum: 15
          maximum: 300
          default: 60
        target_audience:
          type: string
          example: "tech enthusiasts"
        keywords:
          type: array
          items:
            type: string
          example: ["AI", "productivity", "automation"]
        tone:
          type: string
          enum: [casual, formal, enthusiastic, educational]
          default: casual
    ScriptGenerationResponse:
      type: object
      properties:
        script_id:
          type: string
          format: uuid
        content:
          type: string
          description: Generated script content
        word_count:
          type: integer
        estimated_duration:
          type: integer
          description: Estimated reading time in seconds
        style:
          type: string
        status:
          type: string
          enum: [generated, approved, rejected]
        generation_time_seconds:
          type: number
        metadata:
          type: object
    ImageGenerationRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          description: Text description of the desired image
          example: "Modern workspace with AI hologram"
        style:
          type: string
          enum: [modern, vintage, minimalist, artistic, photorealistic, cartoon, anime, cyberpunk]
          default: modern
        aspect_ratio:
          type: string
          enum: ["9:16", "16:9", "1:1", "4:3"]
          default: "9:16"
        resolution:
          type: string
          enum: [720p, 1080p, 4k]
          default: 1080p
        negative_prompt:
          type: string
          description: Elements to avoid in the image
        seed:
          type: integer
          description: Seed for reproducible results
        guidance_scale:
          type: number
          minimum: 1.0
          maximum: 20.0
          default: 7.5
        num_inference_steps:
          type: integer
          minimum: 10
          maximum: 100
          default: 50
    ImageGenerationResponse:
      type: object
      properties:
        image_id:
          type: string
          format: uuid
        image_url:
          type: string
          format: uri
          description: URL to the generated image
        prompt:
          type: string
        style:
          type: string
        aspect_ratio:
          type: string
        resolution:
          type: string
        generation_time:
          type: number
        status:
          type: string
          enum: [completed, failed]
        metadata:
          type: object
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "rate_limit_exceeded"
            message: "Too many requests. Please try again later."
...